Cambio de campos float a decimal(20,2) / Solucionado el error campo id no encontrado al intentar ver los detalles de un usuario / Agregando funcion para calcular el total de las ventas del día y las ventas filtradas (en proceso)

+++ database/migrations/2019_10_10_215020_create_articulos_table.php
@@ -18,9 +18,9 @@ class CreateArticulosTable extends Migration
             $table->string('Nombre');
             $table->unsignedBigInteger('Categoria_Id');
             $table->foreign('Categoria_Id')->references('id')->on('categorias');
-            $table->float('Precio_Costo');
-            $table->float('Precio_Venta');
-            $table->float('Cantidad');
+            $table->decimal('Precio_Costo', 20, 2);
+            $table->decimal('Precio_Venta', 20, 2);
+            $table->decimal('Cantidad', 20, 2);
             $table->boolean('Apto_Venta_Unitaria');
             $table->text('Descripcion');
             $table->string('Campo_Extra_1');
