+++ database/migrations/2019_10_10_221146_create_cajas_table.php
@@ -19,8 +19,8 @@ class CreateCajasTable extends Migration
             $table->string('Forma_Cobro');
             $table->enum('Estado', ['ABIERTA','CERRADA']);
             $table->string('Terminal')->unique()->nullable();
-            $table->float('Monto_Inicial')->nullable();
-            $table->float('Monto_Final')->nullable();
+            $table->decimal('Monto_Inicial', 20, 2)->nullable();
+            $table->decimal('Monto_Final', 20, 2)->nullable();
             $table->text('Descripcion')->nullable();
             $table->dateTime('Fecha_Hora_Apertura')->nullable();
             $table->dateTime('Fecha_Hora_Cierre')->nullable();
